# deprecated/openapi.yaml
openapi: 3.0.1
info:
  title: Transfer Service API
  description: Defintions of Transfer Service End-points
  version: '0.1'
servers:
  - url: 'http://data.uvadcos.io'
paths:

  /:
    get:
      description: Upload Homepage
      responses:
        '200':
          description: HTML Upload Page
      servers:
        - url: 'http://data.uvadcos.io'
    servers:
      - url: 'http://data.uvadcos.io'

  /data/{ark}:
    get:
      summary: Returns file associated with given Identifier
      parameters:
        - in: path
          name: ark
          schema:
            type: string
          required: true
          description: Minted Identifier of desired
      responses:
        '200':
          description: File downloaded
        '400':
          description: "Improperly formatted ID/ID not registered"
      servers:
        - url: 'http://data.uvadcos.io'
    post:
      summary: Uploads file with metadata to ORS
      description: File Upload End-point
      parameters:
        - in: path
          name: ark
          schema:
            type: string
          required: false
          description: Desired Identifier
      responses:
        '200':
          description: File uploaded successfully and minted new id
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                metadata:
                  type: string
                  format: binary
    delete:
      summary: Deletes an object from minio
      parameters:
        - in: path
          name: ark
          schema:
            type: string
          required: true
          description: Identifer of object
      responses:
        '200':
          description: File Deleted
        '400':
          description: Improperly formatted ID/ID not registered
      servers:
        - url: 'http://data.uvadcos.io'
    servers:
      - url: 'http://data.uvadcos.io'

  /bucket/{bucketName}/:
    get:
      summary: Create a bucket in minio
      parameters:
        - in: path
          name: bucketName
          schema:
            type: string
          required: true
          description: Name of new bucket
      responses:
        '200':
          description: File downloaded
        '400':
          description: "Bucket already exists"
      servers:
        - url: 'http://data.uvadcos.io'
    servers:
      - url: 'http://data.uvadcos.io'
    delete:
      summary: Deletes a bucket in minio
      parameters:
        - in: path
          name: bucketName
          schema:
            type: string
          required: true
          description: Name of bucket to be deleted
      responses:
        '200':
          description: Bucket Deleted
        '400':
          description: "Bucket does not exist"
      servers:
        - url: 'http://data.uvadcos.io'

  /run-job/:
    post:
      summary: Runs Spark job on given dataset
      description: Run Job using spark using dataset and software ID's
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Dataset Identifier:
                  type: string
                Job Identifier:
                  type: string
            examples:
              '0':
                value: |
                  {
                      "Job Identifier":"ark:99999/54c75a4e-850b-47a0-a638-19e647f2a1eb",
                      "Dataset Identifier":"ark:99999/6521e5fd-5ce7-4681-b35e-dd575c7c293a"
                  }
      responses:
        '200':
          description: Job Run Successfully
      servers:
        - url: 'http://data.uvadcos.io'
    servers:
      - url: 'http://data.uvadcos.io'

  /download-html/:
    get:
      summary: HTML Download Page
      description: HTML Page to download file for Identifier
      responses:
        '200':
          description: HTML Download Page
      servers:
        - url: 'http://data.uvadcos.io'
    servers:
      - url: 'http://data.uvadcos.io'